requirements.txt
aiohttp
tqdm
watchdog
instaloader
TikTokApi
python-dotenv